requirements-sc.txt
numpy
matplotlib
tensorflow_datasets
boto3
sagemaker